feat: created hot comment api
